backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
spacy>=3.7.0
nltk>=3.8.0
sentence-transformers>=2.2.0
numpy>=1.24.0
# PERFORMANCE ARCHITECTURE: FAISS for ANN-based ultra-fast resume matching
faiss-cpu>=1.7.4
